Added include for live chat system
The live chat system is run on an external node.js server not part of this repository, so I will send everyone the files soon. Currently, there is an issue where socket.io is not found. It was solved [here](https://stackoverflow.com/questions/46151327/keep-getting-socket-io-eio-3transport-pollingt-lvm1sgo-404-errors-when-i-de), so I will continue to try to fix this.