Notifications menu works for now
It doesn't update in real time, because PHP only can retrieve the state of the database when the page was loaded. 

This is also why i had to use browser refresh in some of the admin side's tables...

Will need to consider using an external system to poll the database in an interval to show the counter as live.